+++ app-backend/lambda-overviews/src/main/scala/com/rasterfoundry/lambda/overviews/OverviewGenerator.scala
@@ -16,7 +16,7 @@ import com.rasterfoundry.datamodel.OverviewInput
 object OverviewGenerator extends LazyLogging {
 
   def createProjectOverview(sceneURIs: List[String],
-                            pixelSize: Int): Option[MultibandGeoTiff] = {
+                            pixelSize: Double): Option[MultibandGeoTiff] = {
 
     val rasterSources = sceneURIs.map {
       GeoTiffRasterSource(_).reproject(WebMercator)
@@ -82,9 +82,13 @@ object OverviewGenerator extends LazyLogging {
     )
 
     println("Creating project overview")
+    // James figured out the formula by fitting the plot of
+    // zoom level versus pixel size in meters
+    // 156412 is the pixel size on zoom level 0
     val projectOverviewOption =
-      OverviewGenerator.createProjectOverview(initialProjectScenes,
-                                              overviewInput.pixelSizeMeters)
+      OverviewGenerator.createProjectOverview(
+        initialProjectScenes,
+        156412 / math.pow(2, overviewInput.minZoomLevel))
 
     println(
       "Checking if scenes have been updated or removed from project layer")
@@ -97,11 +101,8 @@ object OverviewGenerator extends LazyLogging {
       case (true, Some(projectOverview)) =>
         writeOverviewToS3(projectOverview, overviewInput.outputLocation)
         println("Updating project layer in API with overview")
-        val layerUpdateStatus = HttpClient.updateProjectWithOverview(
-          authToken,
-          overviewInput.projectId,
-          overviewInput.projectLayerId,
-          overviewInput.outputLocation)
+        val layerUpdateStatus =
+          HttpClient.updateProjectWithOverview(authToken, overviewInput)
         Some(layerUpdateStatus)
       case _ =>
         println(